feat: remember image build history
Saves the image, type, and folder for every image build. For today,
this just lets us remember the history of image build folders and
default to the same folder the next time you build.

Why save more than the last folder, and why save the image name and
type too? There is no registry of past disk images (like we have
with container images, or containers), e.g. they aren't saved to an
OCI registry, and although there. There's some initial discussion,
but no timeline to do this.

Saving this data means we have a list of all past images and at
least some options for providing actions on them, e.g. launching
vfkit from the (container) image instead of just from the ephemeral
container that was used to build it. Including the data in this PR
means we don't need to deal with an imminent breaking format change.

Signed-off-by: Tim deBoer <git@tdeboer.ca>