nit: add rootful / broken image back
### What does this PR do?

Cache broke and was unable to load the image correctly.

Adding docs/img to the directory for containing any documentation and
image assets (not videos).

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A, preview the README.md

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

N/A, preview the README.md

<!-- Please explain steps to reproduce -->

Signed-off-by: Charlie Drage <charlie@charliedrage.com>