chore: CI shouldn't use latest
Our CI shouldn't push the latest changes on main to :latest, because
then anyone who pulls our latest container image will get the
latest, unreleased code. Ideally, they should get the 'latest
release'.

This changes the CI to push to :nightly. For the record I don't love
the term 'nightly' and I don't think it's quite correct, but it is
what AI Studio is using and I think it's better to be consistent. :)

Fixes #96.

Signed-off-by: Tim deBoer <git@tdeboer.ca>