chore: update release version
The 'tag' release job was setting the correct version in package.json,
then bumping it up to create a PR and doing the build, which means the
output container has the bumped version and not the release version.

Looked at other options, but the tasks don't appear to depend on each
other so I'm just swapping the order.

Fixes #116.

Signed-off-by: Tim deBoer <git@tdeboer.ca>