fix: pick unused container name
Make sure we pick an unused container name so that it's possible to
kick off multiple builds at the same time.

Note that there are still two theoretical places this could still fail
due to the delay between when you kick off the task and we check for
conflicts vs when the container gets created: the first time you ever
run (if the bootc builder image takes time to pull down) and if you're
super-fast and run the action twice before the container engine can
create a container and register back in Podman Desktop. The first seems
very unlikely that someone would try, and the second I couldn't even
reproduce when trying.

Fixes #103.

Signed-off-by: Tim deBoer <git@tdeboer.ca>