feat: remove old builder images
Adds a new function to delete images with a given name that do not
have the same tag (nor a second tag; you can tag to keep something).
Uses this to look for and clean up any old builder images, so that
we'll automatically clean up whenever we pull a new one.

Fixes #98.

Signed-off-by: Tim deBoer <git@tdeboer.ca>