docs: update the containerfile for quadlet
### What does this PR do?

Using "pull" no longer works and requires a custom storage.conf

Instead, just remove the "pull" part as the OS will automatically try
and pull the image anyways on boot if it's not available.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

<!-- Please explain steps to reproduce -->

N/A

Signed-off-by: Charlie Drage <charlie@charliedrage.com>