feat: refactor of build logic / more history info / quickpicks (#175)
### What does this PR do?

In order to add the "quickpick" logic in a separate file, it was
beneficial to also refactor what history information we also save.

In this PR we:
* Separate the "quickpick" functionality into a separate file, so we may
  use the build disk image function separately
* Additional history information being saved such as build container id,
  timestamp, etc.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A, internal only. Build functionality should work as intended.

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes #173
Closes #174

### How to test this PR?

<!-- Please explain steps to reproduce -->

Build a disk image, it should continue as normal.

Check:
`~/.local/share/containers/podman-desktop/extensions-storage/redhat.bootc/history.json`
for the 'updated history info'

Signed-off-by: Charlie Drage <charlie@charliedrage.com>