feature: add welcome screen with links
### What does this PR do?

Adds more information to the "blank" screen when there are no builds.
This is intended to get the user to go through the documentation which
is (usually) a one step process / point them in the right direction
of the README, examples, etc.

PR is not meant to be a "step by step" welcome screen, but more of an
initial screen.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes #150

### How to test this PR?

<!-- Please explain steps to reproduce -->

View the webview

Signed-off-by: Charlie Drage <charlie@charliedrage.com>