chore: fix end status (#233)
Fixing two minor issues related to the end of the build task:

- The bug in PD that required us to report an increment of -1 to trigger task completion
appears to have been fixed a while ago. Better to report 100%.

- Currently, the dialog to tell you that the job is successful blocks the task from
finishing until you say OK. It should be async.

I fixed the second by combining the two paths of updating the history into one and
moving both success and failure dialogs to a .then(), but unfortunately that's enough
to tip yarn format to change everything. All of the changes are below line 231.

Fixes #232.

Signed-off-by: Tim deBoer <git@tdeboer.ca>