chore: build multiple image types (and no target arch api) (#223)
* chore: build multiple image types (and no target arch api)

Provide the ability to specify multiple image types at a time. Changed the build
type from a string to a typed enum array and propagated the change everywhere.

Changing to an enum allowed me to remove the 'Invalid image format' error message,
and for now I changed it to loop over all types to check for any existing files
for the warning. Changed the radio buttons in the UI to checkboxes to allow
multiple selection.

I removed the 'bootc.build.image.location' and 'bootc.build.type' labels from the
container (and imagePath parameter to createBuilderImageOptions) since we don't
use them, and they don't make sense/would have to be different when you're building
multiple image types.

Since it was very closely aligned and would create overlapping PRs, I included the
change I started coding this for: changing the API to allow not specifying a target
architecture, which causes the --target-arch parameter to not be passed to the cli.
No support exists in the frontend, this might be provided later as a fix or
temporary workaround for #149.

Fixes #216.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: PR feedback

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>
Signed-off-by: Vladimir Lazar <vlazar@redhat.com>