chore: temporarily ignore arch when building iso (#246)
Image builder has had an issue for a while now where if you specify the
target arch when building an iso it will fail, even if you're specifying the
current/default arch. We've had enough users hit this (so this extension is
the 'blocking' issue) even though the generated images are also
broken/unusable.

We've discussed possible UI changes (like having a 'no target arch' option,
or allowing no arch to be selected) but we don't really like these and do
not want to make UI changes that could be very temporary - so this is the
simplest workaround for now that will unblock users.

Once the following bug is resolved we will update image builder, remove this
workaround, and if necessary change the UI.

https://github.com/osbuild/bootc-image-builder/issues/316

Signed-off-by: Tim deBoer <git@tdeboer.ca>
Signed-off-by: Vladimir Lazar <vlazar@redhat.com>