feat: open webview from image action (#235)
* feat: open webview from image action

Removes the quick pick command, and instead triggers opening the webview, switching to
the build page, and preselecting the image.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: add tests

Adding tests for the command redirecting to the build page, and for the build
page allowing an initial selection.

To test a select control the values must be strings, so I had to rework the
Build page a bit to use string values.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: respond to review comments

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>
Signed-off-by: Vladimir Lazar <vlazar@redhat.com>