feat: add telemetry to pull, goto build page (#288)
* feat: add telemetry to pull, goto build page

Exposed the telemetry logger from the extension, and changed build-disk-image.ts
to use it instead of its own. While doing this I noticed that logUsage was being
called before and after the build, so I removed the first to avoid
double-counting.

Adds telemetry event when pulling images. The intent here is to show how often
users pull the example but will also trigger when pulling the image builder
itself, which might also be useful.

Exposes telemetry logging to frontend, and sends an event when someone clicks
on the button to go to the Build page.

Part of #179.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: telemetryLogUsage and logError

Responding to PR review: rename API to telemetryLogUsage so that it's clear,
and expose logError() while we're at it.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>
Signed-off-by: Vladimir Lazar <vlazar@redhat.com>