chore: expose validation API
Refactoring in preparation for #200 and #262.

Currently, there is no way for the frontend to check prereqs or see if there is
an existing image before attempting to build, at which point the backend has to
open dialogs to notify of an error, or ask the user whether to overwrite.

This PR moves the existing code from the build to check prereqs and for
existing images out into two functions accessible to the frontend:
 - checkPrereq() checks that the machine is Podman v5 and rootful.
 - buildExists() checks if there's an existing build in a given folder.

It also adds an option to the build to overwrite an existing image.

Tests existed for the underlying function and the build just calls the
refactored functions, so there is no behaviour change from before.

Signed-off-by: Tim deBoer <git@tdeboer.ca>