feat: build form validation
Prior to this change, if you were missing prereqs (v5+rootful) or were missing
some option you wouldn't see a problem until you build, and then after you
accepted the error message it would sit on the build screen for 10s before it
timed out. Likewise, if there was an existing build and you said 'no' to
overwriting, it would wait for 10s to time out.

With this change, all validation is done inside the form. If you're missing a
prereq or one of the options, the error message is visible at the bottom of
the form and the Build button is disabled until you fix it. If you need to
confirm overwriting an existing build, a checkbox appears.

Renamed buildErrorMessage so it's more clear this is the post-build error.

Added tests for prereqs and overwriting, and fixed the rather obvious error
in api-impl.ts.

Fixes #262. Fixes #200.

Signed-off-by: Tim deBoer <git@tdeboer.ca>