feature: automatically select image based on image architecture inspect
### What does this PR do?

* Automatically selects the image based upon what architecture is
  available under inspect image.

Note that this is also in anticipation of manifest feature being added,
so there is some code in here (such as an array, rather than a single
string) for manifest.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop-extension-bootc/issues/245

### How to test this PR?

Open a build, try and click on arm64 / amd64.

<!-- Please explain steps to reproduce -->

Signed-off-by: Charlie Drage <charlie@charliedrage.com>