chore: use disk image icon for build
This is minor, but we have a nice icon and the PD standard is more typically
to use the icon for the thing you want to create (i.e. a disk image) not the
thing you have now (container image). We didn't have the 'correct' container
image anyway since it is not exported from PD, yet another reason to use the
disk image icon.

Changes the default disk image icon to be solid and the correct size for
buttons, then uses it for the homepage button, and a larger version for
the Build page.

Signed-off-by: Tim deBoer <git@tdeboer.ca>