chore: remove bib iso target workaround
Bib does not support cross-arch building of ISO images for now:
https://github.com/osbuild/bootc-image-builder/issues/316

However, there was an issue where it would fail if you specify a target arch,
even if that's the current one. This has been fixed and is in the bib that
we currently depend on:
https://github.com/osbuild/bootc-image-builder/pull/364

This removes the workaround we had in place to avoid ISO builds failing.

Fixes #269.

Signed-off-by: Tim deBoer <git@tdeboer.ca>