docs: add some simple instructions on post-building usage
### What does this PR do?

Adds some simple explanations on how to use each format.

We should not go into too much detail as there are **many** solutions to
how to run raw, qcow2, etc, and that generic links should be sufficient.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop-extension-bootc/issues/82

### How to test this PR?

<!-- Please explain steps to reproduce -->

N/A

Signed-off-by: Charlie Drage <charlie@charliedrage.com>