chore: adopt Link
Replace our Link with the component from the UI library. This is the first
case where the component is less functional and just a UI component, so I've
created a new bootc Link component (not in the /upstream folder) that wraps
the ui-svelte Link component and supports open actions (external url, tinro
navigation, and now folders).

Fixes #375.

Signed-off-by: Tim deBoer <git@tdeboer.ca>