chore: adopt checkbox component
The temporary checkbox had some minor UI issues, time to use the Checkbox
component from Podman Desktop UI library. There is a limitation that you
cannot bind to a group array with a value that would have made this
particular case a little simpler, but I just used a function that updates
the buildType array.

Fixes #537 (and also required for light mode #540).

Signed-off-by: Tim deBoer <git@tdeboer.ca>