chore: use ellipsis in columns
We did this in Podman Desktop but should do it in bootc - text columns
that can be too wide should use ellipsis.

The new bootc Link component (on PD Link component) supports opening
folders directly, so we can remove the unused import and function. Added
color variable so that ellipsis is grey instead of white.

Fixes #549.

Signed-off-by: Tim deBoer <git@tdeboer.ca>