chore: light mode, mostly for build page
Move up to latest ui-svelte in order to do the following. Requires a recent
Podman Desktop build in order for variables to be defined at runtime.

In the Build form, pick up status colors, switch to use ErrorMessage component,
and pick up other light mode color variables. Uses the same styling as PD for
the tiles, and radio buttons use the checkbox color variables (since PD has no
vars for radio inputs).

Use status colors in bootc status icon for now - we should probably reuse
StatusIcon now, but that's separate work.

After doing this, we can even set the default text color in the index.

Not marking this as 'fixes' since we still need a select component, but this
is most of #540.

Signed-off-by: Tim deBoer <git@tdeboer.ca>