chore: tailwind cleanup
The bootc extension tailwind config was originally copied from Podman Desktop
and contains a lot of unused content. We should remove all of it so stay
right-sized, so that it's obvious which colors are still hard-coded (should
probably be using color variables), and so that we don't add any new dependency
on colors.

Fixes #627.

Signed-off-by: Tim deBoer <git@tdeboer.ca>