feat: details page for disk images
Modifies the build logs to have a proper details page for disk images, similar
to how Podman Desktop handles Images or Containers.

The two tabs are Summary and Build Logs. The summary page is simple for now
but could be expanded; the build logs doesn't really match what we typically
have in Podman Desktop, but it fits in better here than a standalone page.

Fixes #856.

Signed-off-by: Tim deBoer <git@tdeboer.ca>