feat: go directly to build logs on successful build
When a build is started successfully, go directly to the build logs.

Switches to use the store instead of calling bootcClient.listHistoryInfo()
repeatedly, less for performance/reuse and more to ensure that the store
has been updated before switching to the build logs. Otherwise we may try
to switch before the store has been updated.

Noticed a pattern in a few other tests of clicking and then 'await new Promise',
fixed these to await the userEvent while I was here.

Fixes #872.

Signed-off-by: Tim deBoer <git@tdeboer.ca>