chore: always pass empty body to notify message
Confirmed the error isn't impacting function at all, but use an empty
object instead of an empty string to avoid errors in the log.

The code in api-impl.ts had an optional parameter that was always passed,
just make the parameter mandatory to avoid the same problem in the future.

Fixes #914.

Signed-off-by: Tim deBoer <git@tdeboer.ca>