feat: navigation and dashboard
Introduces left navigation for bootc. The initial Dashboard is just the empty screen
that we had before (but now always visible), and Disk Images is the main screen we
had after you had created an image, renamed and with a more typical empty screen.
This sets bootc up for adding additional pages in the future, most likely Samples
and a filtered bootable Images page.

Most page urls had to be updated, and I switched to vi.waitFor() for the tests I
updated.

Most of #886, but not marking as fixed since there is some renaming & cleanup
that I'd like to do on other files but didn't want to complicate this PR further.

Signed-off-by: Tim deBoer <git@tdeboer.ca>