docs: remove $ from readme for easier copy and paste
### What does this PR do?

Using the copy and paste button on GitHub results in including $.

We should remove the `$` from the README (on our podman desktop site, we
instead just ignore the $ at the beginning when copying and pasting).

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

<!-- Please explain steps to reproduce -->

N/A

Signed-off-by: Charlie Drage <charlie@charliedrage.com>