chore: refit terminal after initial loading
### What does this PR do?

`xterm.js` has an odd issue where it will not resize the terminal on
initial loading.

After a long time troubleshooting, there is no ideal fix as we are
loading the terminal within many other divs and it's unable to properly
refit itself.

This is causing issues with the scroll-bar being inaccurate. Even after
adding proper CSS / tailwind values for spacing, it is inaccurate / not
correct.

This adds a reactive statement that refits the terminal after a very
short 10ms delay after logs appear.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/podman-desktop/podman-desktop-extension-bootc/issues/997

### How to test this PR?

<!-- Please explain steps to reproduce -->

1. Load logs
2. No longer see padding on the bottom

Signed-off-by: Charlie Drage <charlie@charliedrage.com>